Add in support for the optional client-side permalink, which doesn't rely on storing data within the STOQS database.  This can be problematic for a few reasons:
1.  The GET parameter generated can be quite long (it is lza compressed then base64 compressed.  This can be outside the maximum allowable URL length for a browser, which would cause the permalink to not be uncompressed correctly (and fail).  There is some exception handling code to try to catch that and alert the user.

2.  The code relies on your window having the same elements as the old window - if it does not (i.e., parameters are missing) then it's possible that the loading will fail as well.

3.  An additional javacsript library is added for LZA encoding within the browser.

The better way to do this is to store permalinks in the database.  To allow for the switching between the two, you can use the "client_side_permalink" setting.  When set to true, permalinks would be fully-encoded in the URL.  When set to false, they are stored in the database (and are 32 characters long.)  Permalink loading happens based on link length, so regardless of the setting, both client side and server side permalinks would be loaded correctly.

If you want, you can create a checkbox on the page, and when checked client side links are generated, when unchecked server side links, etc.