Massive rework changing about 400 lines in 7 files:
Replaced method for plotting data on Temporal and Spatial sections of the UI.
Instead of requiring a single Measured Parameter to be selected in the filter,
the Plot function is now decoupled from the filter operation.  A new "Plot"
column set of radio buttons now controls which Parameter gets plotted.  This
will work for both Measured and Sampled Parameters.  The client code now does
the check for a single platform and radio buttons inherently ensures a single
parameter is selected.

The KML fed to the OpenLayers vector layer now uses the count and stride infrastructure
added months ago.  The response is now limited to about 300 points, which seems
spiffy enough on my laptop.

All of the updates to the sensortracks and flot plot info have been removed and
a new .data() item added on $('#temporalparameterplotinfo') is used for display of
parameter and platform information.