Add x3dTerrains optional parameter to LoadScript and -v option.  The Spatial 3D tab is now shown only if terrain(s) have been loaded with a dictionary assigned to the x3dTerrains parameter in the constructor of the loader.  See the loaders/CANON/loadCANON_march2013.py loader for an example.
An X3D terrain file may be generated from MB-System using the mbm_grd2geovrml macro and a little bit of
editing.  The InstantReality aopt program is also helpful. The Viewpoint parameters can be had by
viewing the terrain in x3dom and using the ?a? key (show all), double-clicking on the mesh and by printing
the current viewpoint params in the console (-> ?d? key) by pressing ?v?.